Fix race in NodeId innitialization in FlakeId proxy
Fixes #16609

Reasoning:
The `nextNodeIdUpdate` timestamp was updated BEFORE the `nodeId` field.
This means when 2 threads were initializing the proxy concurrently
then the 2nd thread could see the timestamp already updated then skip the
IF branch and return a stale version of nodeId - which was not updated yet.

This problem is fixed by updating the timestamp after the nodeId field.

Additionally when the IF branch is not taken we still have to dereference
the nodeId field. Otherwise we would return a stale value we had dereferenced
before the IF branch.